stablehlo: dtype-aware value path — emit int/bool tensors (#321) (#99)
* stablehlo: add ElemKind axis to the value path (all-Real, no behaviour change)

* stablehlo: emit int/bool value tensors + convert at embedding boundaries

mlir_type_of now reads a node's inferred scalar kind (Real/Int/Bool, spec
§03) instead of hardcoding Real (Complex refuses); Node::Lit(Int)/Lit(Bool)
render as i32/i1 value constants (Emitter::int_value_const/bool_value_const),
distinct from the existing i32-hardcoded control-flow int_const/bool_const.
Emitter::convert emits the canonical booleans⊂integers⊂reals embedding at op
boundaries; compare's result is now Bool.

Kind-polymorphic ops (add/sub/mul/neg/abs) reconcile a mismatched-elem-kind
operand pair inside Emitter::binary/unary itself (mirroring the existing
shape auto-broadcast), not only at ops.rs's dispatch table: crate::registry's
distribution builders call these primitives directly with no NodeId to
coerce against an inferred kind (e.g. a discrete distribution's Int-typed
observed count mixed with a Real parameter — Bernoulli/Poisson/Binomial/
Geometric/NegativeBinomial(2)/Multinomial all do this). Real-only ops
(divide/pow/log/exp/sqrt/cos/invlogit/lgamma) force their operand(s) to Real
first for the same reason — caught via Gamma(2, 1)'s literal integer
shape/rate reaching gamma_logpdf's log/lgamma directly (any literal-parameter
continuous distribution, e.g. Normal(0, 1), hits this). Shape-only ops
(slice/reshape/broadcast_in_dim/transpose/vector's concatenate) now preserve
the operand's elem kind instead of hardcoding Real, since StableHLO requires
these to match exactly (an Int array's get/get0 slice must stay Int).

ops.rs's real-domain heads (divide/pow/log/exp/sqrt/cos/invlogit/sum/
logsumexp/in) convert operands to Real; vector converts each element to the
array's own inferred kind. div/mod are still unimplemented (Task B1),
refusing via the catch-all.

Existing goldens (Bernoulli/Poisson/Binomial/Geometric/NegativeBinomial(2)/
Categorical/Categorical0/Multinomial) re-baselined: only i32/i1 type changes
and inserted convert lines, no arithmetic/numeric drift (checked per-golden).
poisson-model.flatppl's density re-verified end to end: the emitted op
sequence is isomorphic to pre-change modulo the inserted converts, every
dense<> constant is numerically identical, and the module parses under jax's
MLIR parser and executes to a bit-identical value under Enzyme-JAX
(-14.692689895629883, both before and after).

* stablehlo: reconcile element kinds in compare/select/reduce_axis

Task A2 review found these three ops were missed by the self-reconcile
design binary/unary already got: compare/select rendered their operands'
elem kinds correctly but didn't first reconcile a MISMATCHED pair (e.g.
ops::lower_in's compare(int_product, real_zero), an all-integer in(k,
interval(0, 10))), and select/reduce_axis both hardcoded their result elem
to Real regardless of their operands' actual kind (an int ifelse's select,
or sum over an int array's reduce, would emit an int operand against a
mismatched f32 result/init — invalid, since StableHLO requires these to
agree).

compare/select now widen a mismatched operand pair via Emitter::convert
first, same elem_rank order as binary (booleans ⊂ integers ⊂ reals); select
and reduce_axis's result elem now copies the (reconciled) operand's elem
instead of hardcoding Real. reduce_axis's init constant also needs its own
identity literal per elem kind ("0"/"false", not the float-only
"0.000000e+00" reduce_sum passes) — reduce_max's dtype-exact -inf bit
pattern has no non-Real form, but that's never reached with a non-Real
operand in practice (logsumexp's vector argument is always Real by
construction, and diag's row-sum masks against a Real zero via select's own
widening), so it only needs to hold for the additive identity, asserted
defensively.

Widening-only, so every existing all-Real caller (every compare/select/sum/
logsumexp in the existing suite) converts nothing and stays byte-identical
— confirmed by diffing: no golden .mlir file changed, only the 3 new ones
added. New goldens (hand-built Modules, no dependency on the determinizer's
const-fold or modes.rs's free-parameter binding — kept deliberately out of
scope here) parse-validated through jax's MLIR parser: sum over an int
array reduces entirely in i32; an all-integer in(k, interval(0, 10))
widens its product to f32 before comparing against the real zero constant;
an int ifelse emits a fully i32 select.

+++ crates/stablehlo/tests/goldens/categorical0_logdensity.mlir
@@ -1,6 +1,6 @@
 module {
   func.func @logdensity() -> tensor<f32> {
-    %0 = stablehlo.constant dense<1.0> : tensor<f32>
+    %0 = stablehlo.constant dense<1> : tensor<i32>
     %1 = stablehlo.constant dense<0.2> : tensor<f32>
     %2 = stablehlo.constant dense<0.3> : tensor<f32>
     %3 = stablehlo.constant dense<0.5> : tensor<f32>
